# Wavelet/docs/swagger.yaml
basePath: /
definitions:
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
  update.LatestReleaseView:
    properties:
      body:
        type: string
      channel:
        type: string
      current_version:
        type: string
      has_update:
        type: boolean
      html_url:
        type: string
      in_progress:
        type: boolean
      prerelease:
        type: boolean
      published_at:
        type: string
      tag_name:
        type: string
      upgrade_logs:
        items:
          $ref: '#/definitions/update.UpgradeLogRecord'
        type: array
      upgrade_status:
        type: string
      upgrade_supported:
        type: boolean
    type: object
  update.StreamSnapshot:
    properties:
      in_progress:
        type: boolean
      upgrade_logs:
        items:
          $ref: '#/definitions/update.UpgradeLogRecord'
        type: array
      upgrade_status:
        type: string
    type: object
  update.UpgradeLogRecord:
    properties:
      created_at:
        type: string
      level:
        type: string
      message:
        type: string
    type: object
  update.upgradeRequest:
    properties:
      channel:
        type: string
    type: object
...
info:
  contact:
    name: OpenFlare
    url: https://github.com/Rain-kl/OpenFlare
  description: OpenFlare 平台后端 API，提供用户认证、系统配置、任务调度与边缘节点管理能力。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenFlare API
  version: 1.0.0
paths:
...
  /api/v1/openflare/update/latest-release:
    get:
      description: 查询 OpenFlare 服务端最新 GitHub Release 及升级状态，需要管理员权限
      parameters:
      - description: 发布渠道（stable/preview）
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 最新发布信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/update.LatestReleaseView'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取最新服务端发布版本
      tags:
      - openflare-update
  /api/v1/openflare/update/logs/ws:
    get:
      description: 通过 WebSocket 推送升级进度快照，需要管理员权限
      responses:
        "101":
          description: WebSocket 升级日志流
          schema:
            $ref: '#/definitions/update.StreamSnapshot'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 流式获取服务端升级日志
      tags:
      - openflare-update
  /api/v1/openflare/update/manual-upgrade:
    post:
      consumes:
      - application/json
      description: 确认并执行手动上传的服务端升级（当前功能已禁用），需要管理员权限
      produces:
      - application/json
      responses:
        "400":
          description: 功能已禁用或参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 确认手动服务端升级
      tags:
      - openflare-update
  /api/v1/openflare/update/manual-upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传服务端二进制以进行手动升级（当前功能已禁用），需要管理员权限
      parameters:
      - description: 服务端二进制文件
        in: formData
        name: binary
        required: true
        type: file
      produces:
      - application/json
      responses:
        "400":
          description: 功能已禁用或参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 上传手动升级二进制
      tags:
      - openflare-update
  /api/v1/openflare/update/upgrade:
    post:
      consumes:
      - application/json
      description: 从最新 Release 调度 OpenFlare 服务端自动升级，需要管理员权限
      parameters:
      - description: 升级参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/update.upgradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 升级任务已调度
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/update.LatestReleaseView'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 触发服务端自动升级
      tags:
      - openflare-update
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"